Corrected solution
Hi Gayle,

Awesome book. I've learned a ton so far reading it. It has made me think about things differently and made me write cleaner, more performant code. Anyway, below is my issue with your solution:

In your book in Chapter 4 on page 101 you define a Binary Search Tree as "a binary tree in which every node fits a specific ordering property: all left descendents <= n < all right descendents. This must be true for each node n."
And you go on to say, a couple paragraphs below "Note that this inequality must be true for all of a node's descendents, not just its immediate children."

I believe that your solution violates the above definition.

In my proposed file change code, I created a test case that is a valid BST and which evaluates to true in both my Question class and in yours.

Looking at my test case, let's say node 10 is the "n" in your definition. That means ALL descendants of the 10 node's right branch must be greater than 10, right? But if you replace the 11 with a 10 (or the 5 with a 4...or the 9 with an 8), then your Question class will (incorrectly) return true, whereas my solution will return false.

Let me know if you don't agree or if my solution misses any test cases.
It's also fine if you don't use my solution and instead just take care of this problem in a different way in your solution. As long as you correct it, then that's fine with me.

Thanks!