Create QuestionE
This solution is linear. I did not found any bug in it. I would appreciate if you let me know if you found.
The idea is simple, saying checking the elements in the array one by one and if an element found previously found then ignoring the previous one as it does not help to find the shortest sub-sequence and that is it.