Update Question.java
Since nums=[3,1,1], x=1 is also a valid test case for this problem. But currently, the output of the code is -1 as  condition (a[left]==a[mid]) is false and there is not condition for (a[mid]==a[right]).
So, I have removed this condition and it works fine in Leetcode.